Dependancy tidyup and and small tweaks and improvements
